Moved legacy notification_dao
